refactor(storage): isolate secure line streaming
